Adapt preview and stories to the new component structure
