--- test/components/stories/button_component_stories.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-class ButtonComponentStories < ViewComponent::Storybook::Stories
-  story(:default) do
-    controls do
-      select(:type, %w[primary secondary danger link], 'primary')
-    end
-  end
-
-  story(:primary) {}
-  story(:secondary) {}
-  story(:danger) {}
-  story(:link) {}
-end
